+++ lib/src/publication/remote.dart
@@ -17,7 +17,6 @@ import 'dart:math';
 
 import 'package:flutter/widgets.dart';
 
-import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 
 import '../core/signal_client.dart';
@@ -155,9 +154,9 @@ class RemoteTrackPublication<T extends RemoteTrack>
     // filter visible build contexts
     final viewSizes = videoTrack.viewKeys
         .map((e) => e.currentContext)
-        .whereNotNull()
+        .nonNulls
         .map((e) => e.findRenderObject() as RenderBox?)
-        .whereNotNull()
+        .nonNulls
         .where((e) => e.hasSize)
         .map((e) => e.size);
 
@@ -213,7 +212,7 @@ class RemoteTrackPublication<T extends RemoteTrack>
         );
 
         newValue.onVideoViewBuild = (_) {
-          logger.fine('[Visibility] VideoView did build');
+          logger.finer('[Visibility] VideoView did build');
           if (_lastSentTrackSettings?.disabled == true) {
             // quick enable
             _cancelPendingTrackSettingsUpdateRequest?.call();