+++ lib/src/track/local/local.dart
@@ -16,6 +16,7 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'package:flutter_webrtc/flutter_webrtc.dart' as rtc;
 import 'package:meta/meta.dart';
@@ -26,9 +27,11 @@ import '../../extensions.dart';
 import '../../internal/events.dart';
 import '../../logger.dart';
 import '../../participant/remote.dart';
+import '../../support/native.dart';
 import '../../support/platform.dart';
 import '../../types/other.dart';
 import '../options.dart';
+import '../processor.dart';
 import '../remote/audio.dart';
 import '../remote/video.dart';
 import '../track.dart';
@@ -57,7 +60,58 @@ mixin VideoTrack on Track {
 }
 
 /// Used to group [LocalAudioTrack] and [RemoteAudioTrack].
-mixin AudioTrack on Track {}
+mixin AudioTrack on Track {
+  EventChannel? _eventChannel;
+  StreamSubscription? _streamSubscription;
+
+  @override
+  Future<void> onStarted() async {
+    if (enableVisualizer == true) {
+      await startVisualizer();
+    }
+  }
+
+  @override
+  Future<void> onStopped() async {
+    if (enableVisualizer == true) {
+      await stopVisualizer();
+    }
+  }
+
+  Future<void> startVisualizer() async {
+    if (_eventChannel != null) {
+      return;
+    }
+
+    await Native.startVisualizer(mediaStreamTrack.id!);
+
+    _eventChannel = EventChannel(
+        'io.livekit.audio.visualizer/eventChannel-${mediaStreamTrack.id}');
+    _streamSubscription =
+        _eventChannel?.receiveBroadcastStream().listen((event) {
+      //logger.fine('[$objectId] visualizer event(${event})');
+      events.emit(AudioVisualizerEvent(
+        track: this,
+        event: event,
+      ));
+    });
+  }
+
+  Future<void> stopVisualizer() async {
+    if (_eventChannel == null) {
+      return;
+    }
+
+    await Native.stopVisualizer(mediaStreamTrack.id!);
+    events.emit(AudioVisualizerEvent(
+      track: this,
+      event: [],
+    ));
+    await _streamSubscription?.cancel();
+    _streamSubscription = null;
+    _eventChannel = null;
+  }
+}
 
 /// Base class for [LocalAudioTrack] and [LocalVideoTrack].
 abstract class LocalTrack extends Track {
@@ -71,16 +125,22 @@ abstract class LocalTrack extends Track {
 
   bool _stopped = false;
 
+  TrackProcessor? _processor;
+
+  TrackProcessor? get processor => _processor;
+
   LocalTrack(
     TrackType kind,
     TrackSource source,
     rtc.MediaStream mediaStream,
-    rtc.MediaStreamTrack mediaStreamTrack,
-  ) : super(
+    rtc.MediaStreamTrack mediaStreamTrack, {
+    bool? enableVisualizer,
+  }) : super(
           kind,
           source,
           mediaStream,
           mediaStreamTrack,
+          enableVisualizer: enableVisualizer,
         ) {
     mediaStreamTrack.onEnded = () {
       logger.fine('MediaStreamTrack.onEnded()');
@@ -203,6 +263,10 @@ abstract class LocalTrack extends Track {
     final newStream = await LocalTrack.createStream(currentOptions);
     final newTrack = newStream.getTracks().first;
 
+    var processor = _processor;
+
+    await stopProcessor();
+
     // replace track on sender
     try {
       await sender?.replaceTrack(newTrack);
@@ -217,6 +281,10 @@ abstract class LocalTrack extends Track {
     // set new stream & track to this object
     updateMediaStreamAndTrack(newStream, newTrack);
 
+    if (processor != null) {
+      await setProcessor(processor);
+    }
+
     // mark as started
     await start();
 
@@ -225,6 +293,50 @@ abstract class LocalTrack extends Track {
       track: this,
       options: currentOptions,
     ));
+  }
+
+  Future<void> setProcessor(TrackProcessor? processor) async {
+    if (processor == null) {
+      return;
+    }
+
+    if (_processor != null) {
+      await stopProcessor();
+    }
+
+    _processor = processor;
+
+    var processorOptions = ProcessorOptions(
+      kind: kind,
+      track: mediaStreamTrack,
+    );
+
+    await _processor!.init(processorOptions);
+
+    logger.fine('processor initialized');
+
+    events.emit(TrackProcessorUpdateEvent(track: this, processor: _processor));
+  }
+
+  @internal
+  Future<void> stopProcessor({bool keepElement = false}) async {
+    if (_processor == null) return;
+
+    logger.fine('stopping processor');
+    await _processor?.destroy();
+    _processor = null;
+
+    if (!keepElement) {
+      // processorElement?.remove();
+      // processorElement = null;
+    }
+
+    // apply original track constraints in case the processor changed them
+    //await this._mediaStreamTrack.applyConstraints(this._constraints);
+    // force re-setting of the mediaStreamTrack on the sender
+    //await this.setMediaStreamTrack(this._mediaStreamTrack, true);
+
+    events.emit(TrackProcessorUpdateEvent(track: this));
   }
 
   @internal