+++ lib/src/track/options.dart
@@ -14,16 +14,22 @@
 
 import 'package:flutter/foundation.dart' show kIsWeb;
 
+import '../support/native.dart';
 import '../support/platform.dart';
 import '../track/local/audio.dart';
 import '../track/local/video.dart';
 import '../types/video_parameters.dart';
+import 'processor.dart';
 
 /// A type that represents front or back of the camera.
 enum CameraPosition {
   front,
   back,
 }
+
+enum CameraFocusMode { auto, locked }
+
+enum CameraExposureMode { auto, locked }
 
 /// Convenience extension for [CameraPosition].
 extension CameraPositionExt on CameraPosition {
@@ -41,20 +47,32 @@ class CameraCaptureOptions extends VideoCaptureOptions {
   /// set to false to only toggle enabled instead of stop/replaceTrack for muting
   final bool stopCameraCaptureOnMute;
 
+  /// The focus mode to use for the camera.
+  final CameraFocusMode focusMode;
+
+  /// The exposure mode to use for the camera.
+  final CameraExposureMode exposureMode;
+
   const CameraCaptureOptions({
     this.cameraPosition = CameraPosition.front,
+    this.focusMode = CameraFocusMode.auto,
+    this.exposureMode = CameraExposureMode.auto,
     String? deviceId,
     double? maxFrameRate,
     VideoParameters params = VideoParametersPresets.h720_169,
     this.stopCameraCaptureOnMute = true,
+    TrackProcessor<VideoProcessorOptions>? processor,
   }) : super(
           params: params,
           deviceId: deviceId,
           maxFrameRate: maxFrameRate,
+          processor: processor,
         );
 
   CameraCaptureOptions.from({required VideoCaptureOptions captureOptions})
       : cameraPosition = CameraPosition.front,
+        focusMode = CameraFocusMode.auto,
+        exposureMode = CameraExposureMode.auto,
         stopCameraCaptureOnMute = true,
         super(
           params: captureOptions.params,
@@ -202,10 +220,14 @@ abstract class VideoCaptureOptions extends LocalTrackOptions {
   // Limit the maximum frameRate of the capture device.
   final double? maxFrameRate;
 
+  /// A processor to apply to the video track.
+  final TrackProcessor<VideoProcessorOptions>? processor;
+
   const VideoCaptureOptions({
     this.params = VideoParametersPresets.h540_169,
     this.deviceId,
     this.maxFrameRate,
+    this.processor,
   });
 
   @override
@@ -247,8 +269,15 @@ class AudioCaptureOptions extends LocalTrackOptions {
   /// Defaults to true.
   final bool typingNoiseDetection;
 
+  /// Attempt to use voiceIsolation option (if supported by the platform)
+  /// Defaults to true.
+  final bool voiceIsolation;
+
   /// set to false to only toggle enabled instead of stop/replaceTrack for muting
   final bool stopAudioCaptureOnMute;
+
+  /// A processor to apply to the audio track.
+  final TrackProcessor<AudioProcessorOptions>? processor;
 
   const AudioCaptureOptions({
     this.deviceId,
@@ -256,31 +285,50 @@ class AudioCaptureOptions extends LocalTrackOptions {
     this.echoCancellation = true,
     this.autoGainControl = true,
     this.highPassFilter = false,
+    this.voiceIsolation = true,
     this.typingNoiseDetection = true,
     this.stopAudioCaptureOnMute = true,
+    this.processor,
   });
 
   @override
   Map<String, dynamic> toMediaConstraintsMap() {
     var constraints = <String, dynamic>{};
 
-    /// in we platform it's not possible to provide optional and mandatory parameters.
-    /// deviceId is a mandatory parameter
-    if (!kIsWeb || (kIsWeb && deviceId == null)) {
+    if (Native.bypassVoiceProcessing) {
       constraints['optional'] = <Map<String, dynamic>>[
-        <String, dynamic>{'echoCancellation': echoCancellation},
-        <String, dynamic>{'noiseSuppression': noiseSuppression},
-        <String, dynamic>{'autoGainControl': autoGainControl},
-        <String, dynamic>{'voiceIsolation': noiseSuppression},
-        <String, dynamic>{'googDAEchoCancellation': echoCancellation},
-        <String, dynamic>{'googEchoCancellation': echoCancellation},
-        <String, dynamic>{'googEchoCancellation2': echoCancellation},
-        <String, dynamic>{'googNoiseSuppression': noiseSuppression},
-        <String, dynamic>{'googNoiseSuppression2': noiseSuppression},
-        <String, dynamic>{'googAutoGainControl': autoGainControl},
-        <String, dynamic>{'googHighpassFilter': highPassFilter},
-        <String, dynamic>{'googTypingNoiseDetection': typingNoiseDetection},
+        <String, dynamic>{'googEchoCancellation': false},
+        <String, dynamic>{'googEchoCancellation2': false},
+        <String, dynamic>{'googNoiseSuppression': false},
+        <String, dynamic>{'googNoiseSuppression2': false},
+        <String, dynamic>{'googAutoGainControl': false},
+        <String, dynamic>{'googHighpassFilter': false},
+        <String, dynamic>{'googTypingNoiseDetection': false},
+        <String, dynamic>{'noiseSuppression': false},
+        <String, dynamic>{'echoCancellation': false},
+        <String, dynamic>{'autoGainControl': false},
+        <String, dynamic>{'voiceIsolation': false},
+        <String, dynamic>{'googDAEchoCancellation': false},
       ];
+    } else {
+      /// in we platform it's not possible to provide optional and mandatory parameters.
+      /// deviceId is a mandatory parameter
+      if (!kIsWeb || (kIsWeb && deviceId == null)) {
+        constraints['optional'] = <Map<String, dynamic>>[
+          <String, dynamic>{'echoCancellation': echoCancellation},
+          <String, dynamic>{'noiseSuppression': noiseSuppression},
+          <String, dynamic>{'autoGainControl': autoGainControl},
+          <String, dynamic>{'voiceIsolation': noiseSuppression},
+          <String, dynamic>{'googDAEchoCancellation': echoCancellation},
+          <String, dynamic>{'googEchoCancellation': echoCancellation},
+          <String, dynamic>{'googEchoCancellation2': echoCancellation},
+          <String, dynamic>{'googNoiseSuppression': noiseSuppression},
+          <String, dynamic>{'googNoiseSuppression2': noiseSuppression},
+          <String, dynamic>{'googAutoGainControl': autoGainControl},
+          <String, dynamic>{'googHighpassFilter': highPassFilter},
+          <String, dynamic>{'googTypingNoiseDetection': typingNoiseDetection},
+        ];
+      }
     }
 
     if (deviceId != null) {